requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.3.4
    # via django
cffi==1.14.5
    # via cryptography
cryptography==3.4.7
    # via pyjwt
django-debug-toolbar==3.2.1
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django==3.2.4
    # via
    #   django-debug-toolbar
    #   django-filter
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.1.0
    # via -r requirements.in
pytimeparse==1.1.8
    # via -r requirements.in
pytz==2021.1
    # via django
shortuuid==1.0.1
    # via -r requirements.in
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
